+++ src/command/render/latexmk/pdf.ts
@@ -14,7 +14,6 @@ import { ProcessResult } from "../../../core/process.ts";
 
 import { hasTexLive } from "./texlive.ts";
 import { runBibEngine, runIndexEngine, runPdfEngine } from "./latex.ts";
-import { LatexmkOptions } from "./latexmk.ts";
 import { PackageManager, packageManager } from "./pkgmgr.ts";
 import {
   findIndexError,
@@ -25,9 +24,8 @@ import {
   needsRecompilation,
 } from "./parse-error.ts";
 import { info, warning } from "log/mod.ts";
+import { kLatexHeaderMessageOptions, LatexmkOptions } from "./types.ts";
 
-export const kLatexHeaderMessageOptions = { bold: true };
-export const kLatexBodyMessageOptions = { indent: 2 };
 export async function generatePdf(mkOptions: LatexmkOptions): Promise<string> {
   if (!mkOptions.quiet) {
     info(
